refactor(iam): kill resolves module :'(
ref: MANAGER-11434
