+++ packages/manager/modules/iam/src/iam.module.js
@@ -8,7 +8,6 @@ import ngOvhFeatureFlipping from '@ovh-ux/ng-ovh-feature-flipping';
 import ngOvhUtils from '@ovh-ux/ng-ovh-utils';
 import uiKit from '@ovh-ux/ui-kit';
 
-import resolves from './resolves';
 import services from './services';
 import components from './components';
 
@@ -16,6 +15,7 @@ import identities from './identities';
 import onboarding from './onboarding';
 import policy from './policy';
 
+import paramTypes from './iam.paramTypes';
 import routing from './iam.routing';
 
 import '@ovh-ux/ui-kit/dist/css/oui.css';
@@ -23,6 +23,17 @@ import 'ovh-ui-kit-bs/dist/css/oui-bs3.css';
 import './iam.styles.scss';
 
 const moduleName = 'ovhManagerIAM';
+
+/**
+ * Register each param type onto the $urlMatcherFactoryProvider
+ * @see https://ui-router.github.io/ng1/docs/latest/interfaces/params.paramdeclaration.html#type
+ * @see https://ui-router.github.io/ng1/docs/latest/interfaces/params.paramtypedefinition.html
+ * @param {UrlMatcherFactoryProvider} $urlMatcherFactoryProvider
+ */
+const registerTypes = /* @ngInject */ ($urlMatcherFactoryProvider) =>
+  Object.values(paramTypes).forEach(({ type, ...definition }) =>
+    $urlMatcherFactoryProvider.type(type, definition),
+  );
 
 angular
   .module(moduleName, [
@@ -32,7 +43,6 @@ angular
     ngOvhFeatureFlipping,
     ngOvhUtils,
     uiKit,
-    resolves,
     services,
     components,
 
@@ -40,6 +50,8 @@ angular
     onboarding,
     policy,
   ])
-  .config(routing);
+  .config(routing)
+  .config(registerTypes)
+  .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
