+++ packages/manager/modules/iam/src/policy/create/create.module.js
@@ -6,6 +6,9 @@ import routing from './create.routing';
 
 const moduleName = 'ovhManagerIAMPolicyCreate';
 
-angular.module(moduleName, [createPolicy]).config(routing);
+angular
+  .module(moduleName, [createPolicy])
+  .config(routing)
+  .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
