+++ packages/manager/modules/iam/src/policy/create/create.routing.js
@@ -1,11 +1,10 @@
-import { createPolicyBreadcrumbResolve } from '../../resolves';
-
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('iam.createPolicy', {
     url: '/policy/create',
     component: 'iamCreatePolicy',
     resolve: {
-      breadcrumb: createPolicyBreadcrumbResolve,
+      breadcrumb: /* @ngInject */ ($translate) =>
+        $translate.instant('iam_policy_create'),
     },
   });
 };
