+++ packages/manager/modules/iam/src/policy/create/translations/Messages_fr_FR.json
@@ -1,0 +1,3 @@
+{
+  "iam_policy_create": "Créer une politique"
+}
