+++ packages/manager/modules/iam/src/policy/edit/translations/Messages_fr_FR.json
@@ -1,0 +1,3 @@
+{
+  "iam_policy_edit": "Modification de la politique {{ policy }}"
+}
