+++ packages/manager/modules/iam/src/policy/policies/delete/delete.routing.js
@@ -1,8 +1,4 @@
-import {
-  entityResolve,
-  policyParamResolve,
-  statementResolve,
-} from '../../../resolves';
+import { ENTITY } from '../../../iam.constants';
 
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('iam.policy.policies.delete', {
@@ -10,11 +6,36 @@ export default /* @ngInject */ ($stateProvider) => {
     component: 'iamDeleteEntity',
     resolve: {
       breadcrumb: () => null,
-      entity: entityResolve,
-      identity: () => null,
-      policy: policyParamResolve,
-      resourceGroup: () => null,
-      statement: statementResolve,
+
+      /**
+       * A polymorphic DTO required by the deleteEntity component
+       * @returns {{
+       *   data: Object,
+       *   type: string
+       * }|null}
+       */
+      entity: /* @ngInject */ (policy) => {
+        if (policy) {
+          return { data: policy, type: ENTITY.POLICY };
+        }
+        return null;
+      },
+
+      /**
+       * The Policy parameter based on the policy's id
+       * @returns {Object|null}
+       */
+      policy: /* @ngInject */ ($transition$, PolicyService) => {
+        const { policy: uuid } = $transition$.params();
+        return uuid ? PolicyService.getPolicy(uuid) : null;
+      },
+
+      /**
+       * Whether the entity requires a statement
+       * @returns {boolean}
+       */
+      statement: /* @ngInject */ (entity) =>
+        [ENTITY.POLICY].includes(entity.type),
     },
   });
 };
