+++ packages/manager/modules/iam/src/policy/policies/policies.routing.js
@@ -1,15 +1,22 @@
-import { cursorsParamResolve } from '../../resolves';
+import { cursorsType } from '../../iam.paramTypes';
 
 export default /* @ngInject */ ($stateProvider) => {
   $stateProvider.state('iam.policy.policies', {
     url: `?cursors`,
     component: 'iamPolicies',
     params: {
-      cursors: cursorsParamResolve.declaration,
+      cursors: {
+        array: false,
+        dynamic: true,
+        inherit: false,
+        squash: true,
+        type: cursorsType,
+        value: null,
+      },
     },
     resolve: {
       breadcrumb: () => null,
-      cursors: cursorsParamResolve,
+      cursors: /* @ngInject */ ($transition$) => $transition$.params().cursors,
     },
     redirectTo: (transition) =>
       transition
