--- packages/manager/modules/iam/src/resolves/guides.resolve.js
@@ -1,16 +1,0 @@
-import { GUIDE } from '../iam.constants';
-
-// ---------------------------------------------------------------------------------------------------- //
-
-/**
- * The on boarding guides
- * @returns {Object}
- */
-const onboardingGuidesResolve = /* @ngInject */ (GuideService) =>
-  GuideService.formatGuides(GUIDE.IAM, GUIDE.USERS, GUIDE.SAMLSSO);
-
-onboardingGuidesResolve.key = 'onboardingGuides';
-
-// ---------------------------------------------------------------------------------------------------- //
-
-export { onboardingGuidesResolve }; // eslint-disable-line
