--- packages/manager/modules/iam/src/resolves/translations/Messages_fr_FR.json
@@ -1,6 +1,0 @@
-{
-  "iam_resolves_breadcrumbs_create_policy": "Créer une politique",
-  "iam_resolves_breadcrumbs_default": "Gestion des identités et des accès (IAM)",
-  "iam_resolves_breadcrumbs_edit_policy": "Modification de la politique {{ policy }}",
-  "iam_resolves_breadcrumbs_policy_identities": "Gérez les identités de la politique {{ policy }}"
-}
