--- packages/manager/modules/iam/src/resolves/types/index.js
@@ -1,12 +1,0 @@
-import cursors from './cursors.type';
-import urn from './urn.type';
-import uuid from './uuid.type';
-
-export * from './cursors.type';
-export * from './urn.type';
-export * from './uuid.type';
-export default {
-  cursors,
-  urn,
-  uuid,
-};
