+++ packages/manager/modules/iam/src/translations/Messages_fr_FR.json
@@ -1,0 +1,3 @@
+{
+  "iam_breadcrumb": "Gestion des identités et des accès (IAM)"
+}
